backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-core
openai
faiss-cpu
tiktoken
python-dotenv
pydantic